Switched to using "airblast functionality flags" attribute instead of DeflectPlayer hook for sm_dodgeball_deflect_players.

+++ extension/CEntity/CEconItemView.h
@@ -54,7 +54,7 @@ class CEconItemAttribute
 public:
 	CEconItemAttribute() {};
 
-	CEconItemAttribute(int iAttributeDefinitionIndex, float flValue)
+	CEconItemAttribute(uint16 iAttributeDefinitionIndex, float flValue)
 	{
 		this->m_iAttributeDefinitionIndex = iAttributeDefinitionIndex;
 		this->m_flValue = flValue;
